Moved PodOverView. Moved and renamed PodDetailsContainer
